claude-code/requirements.txt
# AniLoader – Dependencies
# Python 3.11+

# HTTP client
httpx>=0.27,<1.0

# HTML parsing
beautifulsoup4>=4.12,<5.0

# Web framework
fastapi>=0.110,<1.0
uvicorn[standard]>=0.29,<1.0
python-multipart>=0.0.9,<1.0

# Config
pyyaml>=6.0,<7.0

# CLI tool for actual downloads (installed separately via pip)
# aniworld
